// src/types/settings.ts
// src/types/settings.ts

export type AccountFormData = {
    avatar: string;
    displayName: string;
    username: string;
    fullName: string;
    email: string;
    phone: string;
    bio: string;
    secondaryEmail: string;
    phoneNumber: string;
  };
  
  export type AddressData = {
    name: string;
    address: string;
    city: string;
    state: string;
    zip: string;
    country: string;
  };
  